removed some warnings

+++ src/Tool/IntelliTool.cpp
@@ -57,8 +57,8 @@ void IntelliTool::onWheelScrolled(int value){
 
 bool IntelliTool::createToolLayer(){
 		if(Area->createTempTopLayer(Area->activeLayer)) {
-				this->activeLayer = &Area->layerBundle[static_cast<unsigned long long>(Area->activeLayer)];
-				this->Canvas = &Area->layerBundle[static_cast<unsigned long long>(Area->activeLayer + 1)];
+                this->activeLayer = &Area->layerBundle[static_cast<size_t>(Area->activeLayer)];
+                this->Canvas = &Area->layerBundle[static_cast<size_t>(Area->activeLayer + 1)];
 				return true;
 		}
 		return false;
